requirements.txt
google-cloud-asset
google-cloud-compute
google-cloud-storage
google-cloud-resource-manager
google-api-python-client
pytest
